webserver/requirements.txt
django==3.0.7
psycopg2==2.8.5
djangorestframework==3.10.3
PyJWT==1.7.1
gunicorn==20.0.4
channels==2.4.0
django-rest-swagger==2.2.0
django-extensions==2.2.9
